linting

+++ src/midiconversion/Converter.java
@@ -27,7 +27,7 @@ public class Converter{
         tempoTrack.insertEvent(ts);
         tempoTrack.insertEvent(t);
           
-        double crotchetLength = 60000.0 / (bpm * PPQ);
+        int crotchetLength = 60000 / (bpm * PPQ);
         int nextNote = crotchetLength;
         for(Bar bar : piece){
             for(Chord chord: bar){
